[API] Enable contract tests for Shipping Method resource
