+++ tests/Api/Shop/ShippingMethodsTest.php
@@ -20,27 +20,25 @@ use Sylius\Component\Core\Model\Address;
 use Sylius\Component\Core\Model\ShipmentInterface;
 use Sylius\Component\Core\Repository\ShipmentRepositoryInterface;
 use Sylius\Tests\Api\JsonApiTestCase;
-use Sylius\Tests\Api\Utils\ShopUserLoginTrait;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Messenger\MessageBusInterface;
 
 final class ShippingMethodsTest extends JsonApiTestCase
 {
-    use ShopUserLoginTrait;
+    protected function setUp(): void
+    {
+        $this->setUpDefaultGetHeaders();
+
+        parent::setUp();
+    }
 
     /** @test */
-    public function it_gets_all_available_shipping_methods_by_default_in_given_channel(): void // Cover case of shipping methods in other channel
+    public function it_gets_all_available_shipping_methods_by_default_in_given_channel(): void
     {
         $this->loadFixturesFromFiles(['channel.yaml', 'cart.yaml', 'country.yaml', 'shipping_method.yaml']);
 
-        $this->client->request(
-            method: 'GET',
-            uri: '/api/v2/shop/shipping-methods',
-            server: self::CONTENT_TYPE_HEADER,
-        );
-        $response = $this->client->getResponse();
+        $this->requestGet('/api/v2/shop/shipping-methods');
 
-        $this->assertResponse($response, 'shop/shipping_method/get_shipping_methods_response', Response::HTTP_OK);
+        $this->assertResponse($this->client->getResponse(), 'shop/shipping_method/get_shipping_methods_response');
     }
 
     /** @test */
@@ -49,27 +47,25 @@ final class ShippingMethodsTest extends JsonApiTestCase
         $this->loadFixturesFromFiles(['channel.yaml', 'cart.yaml', 'country.yaml', 'shipping_method.yaml']);
 
         $tokenValue = 'nAWw2jewpA';
-        $customer = 'sylius@example.com';
-        $this->getCartAndPutItemForCustomer($tokenValue, $customer);
+        $this->getCartAndPutItemForCustomer($tokenValue, 'sylius@example.com');
 
-        $this->client->request('GET', '/api/v2/shop/orders/nAWw2jewpA', [], [], self::CONTENT_TYPE_HEADER);
+        $this->requestGet(sprintf('/api/v2/shop/orders/%s', $tokenValue));
         $orderResponse = json_decode($this->client->getResponse()->getContent(), true);
 
-        $this->client->request(
-            method: 'GET',
-            uri: sprintf('/api/v2/shop/shipping-methods?shipmentId=%s&tokenValue=%s', $orderResponse['shipments'][0]['id'], $tokenValue),
-            server: self::CONTENT_TYPE_HEADER,
-        );
-        $response = $this->client->getResponse();
+        $this->requestGet(sprintf(
+            '/api/v2/shop/orders/%s/shipments/%s/methods',
+            $tokenValue,
+            $orderResponse['shipments'][0]['id'],
+        ));
 
-        $this->assertResponse($response, 'shop/shipping_method/get_order_shipping_methods_response', Response::HTTP_OK);
+        $this->assertResponse($this->client->getResponse(), 'shop/shipping_method/get_order_shipping_methods_response');
     }
 
     /** @test */
     public function it_returns_list_of_available_shipping_methods_of_assigned_cart_for_visitor(): void
     {
         $this->loadFixturesFromFiles([
-            'authentication/customer.yaml',
+            'authentication/shop_user.yaml',
             'channel.yaml',
             'cart.yaml',
             'country.yaml',
@@ -77,31 +73,23 @@ final class ShippingMethodsTest extends JsonApiTestCase
         ]);
 
         $tokenValue = 'nAWw2jewpA';
-        $customer = 'oliver@doe.com';
-
-        $this->getCartAndPutItemForCustomer($tokenValue, $customer);
+        $this->getCartAndPutItemForCustomer($tokenValue, 'oliver@doe.com');
 
         /** @var ShipmentRepositoryInterface $shipmentRepository */
         $shipmentRepository = $this->get('sylius.repository.shipment');
-
         /** @var ShipmentInterface $shipment */
         $shipment = $shipmentRepository->findOneBy([]);
 
-        $this->client->request(
-            method: 'GET',
-            uri: sprintf('/api/v2/shop/shipping-methods?shipmentId=%s&tokenValue=nAWw2jewpA', $shipment->getId()),
-            server: self::CONTENT_TYPE_HEADER,
-        );
-        $response = $this->client->getResponse();
+        $this->requestGet(sprintf('/api/v2/shop/orders/%s/shipments/%s/methods', $tokenValue, $shipment->getId()));
 
-        $this->assertResponse($response, 'shop/shipping_method/get_order_shipping_methods_response', Response::HTTP_OK);
+        $this->assertResponse($this->client->getResponse(), 'shop/shipping_method/get_order_shipping_methods_response');
     }
 
     /** @test */
     public function it_returns_list_of_available_shipping_methods_of_assigned_cart_for_other_users_if_shipment_id_and_cart_provided(): void
     {
         $this->loadFixturesFromFiles([
-            'authentication/customer.yaml',
+            'authentication/shop_user.yaml',
             'channel.yaml',
             'cart.yaml',
             'country.yaml',
@@ -109,10 +97,7 @@ final class ShippingMethodsTest extends JsonApiTestCase
         ]);
 
         $tokenValue = 'nAWw2jewpA';
-        $customer = 'oliver@doe.com';
-        $otherCustomer = 'dave@doe.com';
-
-        $this->getCartAndPutItemForCustomer($tokenValue, $customer);
+        $this->getCartAndPutItemForCustomer($tokenValue, 'oliver@doe.com');
 
         /** @var ShipmentRepositoryInterface $shipmentRepository */
         $shipmentRepository = $this->get('sylius.repository.shipment');
@@ -120,66 +105,52 @@ final class ShippingMethodsTest extends JsonApiTestCase
         /** @var ShipmentInterface $shipment */
         $shipment = $shipmentRepository->findOneBy([]);
 
-        $this->logInShopUser($otherCustomer);
-
-        $this->client->request(
-            method: 'GET',
-            uri: sprintf('/api/v2/shop/shipping-methods?shipmentId=%s&tokenValue=nAWw2jewpA', $shipment->getId()),
-            server: self::CONTENT_TYPE_HEADER,
+        $this->requestGet(
+            uri: sprintf('/api/v2/shop/orders/%s/shipments/%s/methods', $tokenValue, $shipment->getId()),
+            headers: $this->headerBuilder()->withShopUserAuthorization('dave@doe.com')->build(),
         );
-        $response = $this->client->getResponse();
 
-        $this->assertResponse($response, 'shop/shipping_method/get_order_shipping_methods_response', Response::HTTP_OK);
+        $this->assertResponse($this->client->getResponse(), 'shop/shipping_method/get_order_shipping_methods_response');
     }
 
     /** @test */
     public function it_returns_empty_list_of_available_shipping_methods_for_not_existent_shipment(): void
     {
         $this->loadFixturesFromFiles([
-            'authentication/customer.yaml',
+            'authentication/shop_user.yaml',
             'channel.yaml',
             'cart.yaml',
             'country.yaml',
             'shipping_method.yaml',
         ]);
 
-        $this->client->request(
-            method: 'GET',
-            uri: '/api/v2/shop/shipping-methods?shipmentId=-10&tokenValue=nAWw2jewpA',
-            server: self::CONTENT_TYPE_HEADER,
-        );
-        $response = $this->client->getResponse();
+        $this->requestGet('/api/v2/shop/orders/nAWw2jewpA/shipments/-10/methods');
 
-        $this->assertResponse($response, 'shop/shipping_method/get_empty_order_shipping_methods_response', Response::HTTP_OK);
+        $this->assertResponse($this->client->getResponse(), 'shop/shipping_method/get_empty_order_shipping_methods_response');
     }
 
     /** @test */
     public function it_returns_empty_list_of_available_shipping_methods_for_not_existent_order(): void
     {
         $this->loadFixturesFromFiles([
-            'authentication/customer.yaml',
+            'authentication/shop_user.yaml',
             'channel.yaml',
             'cart.yaml',
             'country.yaml',
             'shipping_method.yaml',
         ]);
 
-        $this->client->request(
-            method: 'GET',
-            uri: '/api/v2/shop/shipping-methods?shipmentId=-10&tokenValue=test',
-            server: self::CONTENT_TYPE_HEADER,
-        );
-        $response = $this->client->getResponse();
+        $this->requestGet('/api/v2/shop/orders/test/shipments/-10/methods');
 
-        $this->assertResponse($response, 'shop/shipping_method/get_empty_order_shipping_methods_response', Response::HTTP_OK);
+        $this->assertResponse($this->client->getResponse(), 'shop/shipping_method/get_empty_order_shipping_methods_response');
     }
 
-    private function getCartAndPutItemForCustomer(string $tokenValue, string $customer): void
+    private function getCartAndPutItemForCustomer(string $tokenValue, string $customerEmail): void
     {
         /** @var MessageBusInterface $commandBus */
         $commandBus = self::getContainer()->get('sylius.command_bus');
 
-        $pickupCartCommand = new PickupCart($tokenValue, 'en_US');
+        $pickupCartCommand = new PickupCart($tokenValue);
         $pickupCartCommand->setChannelCode('WEB');
         $commandBus->dispatch($pickupCartCommand);
 
@@ -195,7 +166,7 @@ final class ShippingMethodsTest extends JsonApiTestCase
         $address->setCountryCode('US');
         $address->setPostcode('90000');
 
-        $updateCartCommand = new UpdateCart($customer, $address);
+        $updateCartCommand = new UpdateCart($customerEmail, $address);
         $updateCartCommand->setOrderTokenValue($tokenValue);
         $commandBus->dispatch($updateCartCommand);
     }
